requirements.txt
discord.py>=2.4,<3
python-dotenv>=1.0,<2
aiosqlite>=0.20,<1
